CR fixes: extracted InitProviders component in separate file
